[apt][repository] Trust new apt key
Ease transition to new apt key that will sign our repo in the future.

This ensures that the cookbook installs the new key. When we switch to
using the new key to sign the `Release` file of the apt repo, this will
make sure that the new key is trusted.

After the switch, we should make this new key the main key in the
`apt_repository` resource and remove this `execute` resource.
Optionally, we could make the cookbook delete the old key.